CI/CD, documentation, tests, bug fixes, and kernel feature depth
CI/CD infrastructure:
  - Release workflow (.github/workflows/release.yml): tag builds with auto-release
  - Stale bot (.github/stale.yml): 90/14 day stale/close policy
  - Issue templates (bug_report, feature_request) with labels
  - PR template with build/test/docs checklist
  - Dependabot config for Actions + Docker weekly updates
  - CODEOWNERS set to @rusik69

Documentation:
  - CONTRIBUTING.md (308 lines): build guide, coding style, PR process
  - CODE_OF_CONDUCT.md (133 lines): Contributor Covenant v2.1
  - CHANGELOG.md (447 lines): structured changelog with unreleased section
  - README badges: LOC, C files, license, arch, PRs welcome

Static analysis fixes:
  - audit.c: changed audit_sequence from int to unsigned int (4 %u format bugs)
  - fault.c: cast smp_get_cpu_id() to unsigned int for %u format

Kernel features implemented:
  - 6lowpan.c: Full RFC 6282 IPHC decompression (all TF/NH/HLIM/SAC/DAC modes)
  - af_packet.c: Multicast group membership (PACKET_ADD/DROP_MEMBERSHIP)
  - af_packet.h: Added packet_mreq, tpacket_auxdata, mc_entry structures
  - iso9660.c: Relative symlink resolution via Rock Ridge (SL entries)
  - file_lock.c: Blocking wait with 30s timeout via waitqueue

Test expansion:
  - kunit_net.c (16 tests): Socket, TCP, UDP, ARP, netif lifecycle
  - kunit_vfs.c (15 tests): File ops, directory ops, mount, symlinks, permissions
  - kunit_sched.c enhanced (19 tests): Waitqueue, priority, nice, scheduler hooks
  - kunit_tests.c: Registration for new KUnit test suites
  - test_list.c (21 tests): Linked list, hash table with chaining
  - test_string.c (58 tests): All ISO C string functions with edge cases
  - test_bitmap.c (35 tests): find_first_set, hweight, bitmap operations

Build: ZERO errors, ZERO warnings. All 37 host-side unit tests pass.

+++ src/kernel/fault.c
@@ -425,7 +425,7 @@ static void double_fault_handler(struct interrupt_frame *frame) {
         int n = snprintf(msg, sizeof(msg),
             "DOUBLE FAULT at RIP=0x%lx CR2=0x%lx cpu=%u pid=%u",
             (unsigned long)frame->rip, (unsigned long)cr2,
-            smp_get_cpu_id(), proc ? (unsigned int)proc->pid : 0);
+            (unsigned int)smp_get_cpu_id(), proc ? (unsigned int)proc->pid : 0);
         if (n < 0 || n >= (int)sizeof(msg)) {
             /* Truncation is acceptable — msg is best-effort here */
         }
